Move MetaTest internals to execution package
