+++ src/main/java/nl/tudelft/cse1110/andy/execution/metatest/evaluators/LineReplacementEvaluator.java
@@ -1,4 +1,4 @@
-package nl.tudelft.cse1110.andy.config.metatest.evaluators;
+package nl.tudelft.cse1110.andy.execution.metatest.evaluators;
 
 import java.util.ArrayList;
 import java.util.List;
